+++ src/plotly_graphs/plotly_maps.py
@@ -1,5 +1,5 @@
 from bisect import bisect_right
-from typing import Tuple
+from typing import Tuple, Optional
 
 import numpy as np
 import plotly.graph_objects as go
@@ -174,6 +174,7 @@ def plot_dike_traject_reliability_initial_assessment_map(
     # Update layout of the figure and add token for mapbox
     _middle_point = get_middle_point(dike_traject.dike_sections)
     update_layout_map_box(fig, _middle_point)
+    place_legend_right_top_corner(fig)
 
     return fig
 
@@ -418,6 +419,7 @@ def plot_dike_traject_urgency(
     # Update layout of the figure and add token for mapbox
     _middle_point = get_middle_point(dike_traject.dike_sections)
     update_layout_map_box(fig, _middle_point)
+    place_legend_right_top_corner(fig)
 
     return fig
 
@@ -491,12 +493,13 @@ def plot_dike_traject_measures_map(
 
     _middle_point = get_middle_point(dike_traject.dike_sections)
     update_layout_map_box(fig, _middle_point)
+    place_legend_right_top_corner(fig)
 
     return fig
 
 
 def add_measure_type_trace(
-        fig: go.Figure, section: DikeSection, measure_results: dict, legend_display: dict, opacity: float = 1
+        fig: go.Figure, section: DikeSection, measure_results: dict, legend_display: dict, opacity: float = 1, legendgroup: Optional[str] = None
 ):
     """
     This function adds a trace to the figure for the measure type.
@@ -504,6 +507,9 @@ def add_measure_type_trace(
     :param section: DikeSection
     :param measure_results:
     :param legend_display: dict to avoid double legend entries
+    :param opacity: float
+    :param legendgroup: Optional[str]. If None, the legendgroup is based on the measures types (all VZG grouped together)
+    otherwise it is based on the provided name
     """
     if MeasureTypeEnum.SOIL_REINFORCEMENT.name in measure_results[
         "type"] or MeasureTypeEnum.SOIL_REINFORCEMENT.legacy_name in measure_results[
@@ -550,7 +556,9 @@ def add_measure_type_trace(
         fig.add_trace(
             go.Scattermap(
                 name=_name,
-                legendgroup=_name,
+                legendgroup=_name if legendgroup is None else legendgroup,
+                legendgrouptitle_text=None if legendgroup is None else legendgroup,
+                legendgrouptitle = dict(font=dict(weight='bold')),
                 mode="lines",
                 lat=[x[0] for x in _coordinates_wgs],
                 lon=[x[1] for x in _coordinates_wgs],
@@ -573,7 +581,7 @@ def add_measure_type_trace(
         fig.add_trace(
             go.Scattermap(
                 name="Verticale pipingoplossing",
-                legendgroup="VZG",
+                legendgroup="VZG" if legendgroup is None else legendgroup,
                 mode="lines",
                 lat=[x[0] for x in _coordinates_wgs],
                 lon=[x[1] for x in _coordinates_wgs],
@@ -599,7 +607,7 @@ def add_measure_type_trace(
         fig.add_trace(
             go.Scattermap(
                 name="Stabiliteitsscherm",
-                legendgroup="screen",
+                legendgroup="screen" if legendgroup is None else legendgroup,
                 mode="lines",
                 lat=[x[0] for x in _coordinates_wgs],
                 lon=[x[1] for x in _coordinates_wgs],
@@ -624,7 +632,7 @@ def add_measure_type_trace(
         fig.add_trace(
             go.Scattermap(
                 name="Zelfkerende constructie",
-                legendgroup="diaphram wall",
+                legendgroup="diaphram wall" if legendgroup is None else legendgroup,
                 mode="lines",
                 lat=[x[0] for x in _coordinates_wgs],
                 lon=[x[1] for x in _coordinates_wgs],
@@ -650,7 +658,7 @@ def add_measure_type_trace(
         fig.add_trace(
             go.Scattermap(
                 name="Damwandconstructie",
-                legendgroup="sheetpile",
+                legendgroup="sheetpile" if legendgroup is None else legendgroup,
                 mode="lines",
                 lat=[x[0] for x in _coordinates_wgs],
                 lon=[x[1] for x in _coordinates_wgs],
@@ -679,7 +687,7 @@ def add_measure_type_trace(
         fig.add_trace(
             go.Scattermap(
                 name="Aanpassing bekleding",
-                legendgroup="revetment",
+                legendgroup="revetment" if legendgroup is None else legendgroup,
                 mode="lines",
                 lat=[x[0] for x in _coordinates_wgs],
                 lon=[x[1] for x in _coordinates_wgs],
@@ -708,7 +716,7 @@ def add_measure_type_trace(
         fig.add_trace(
             go.Scattermap(
                 name="Custom",
-                legendgroup="custom",
+                legendgroup="custom" if legendgroup is None else legendgroup,
                 mode="lines",
                 lat=[x[0] for x in _coordinates_wgs],
                 lon=[x[1] for x in _coordinates_wgs],
@@ -1008,6 +1016,22 @@ def update_layout_map_box(fig: go.Figure, center: tuple[float, float], zoom: int
         ),
     )
 
+def place_legend_left_top_corner(fig: go.Figure):
+    fig.update_layout(legend=dict(
+        yanchor="top",
+        y=0.99,
+        xanchor="left",
+        x=0.01,
+    ))
+
+def place_legend_right_top_corner(fig: go.Figure):
+    fig.update_layout(legend=dict(
+        yanchor="top",
+        y=0.95,
+        xanchor="right",
+        x=0.99,
+    ))
+
 
 def add_colorscale_bar(
         fig: go.Figure,
@@ -1046,6 +1070,8 @@ def add_colorscale_bar(
                 ticktext=["1e-2", "1e-3", "1e-4", "1e-5", "1e-6"],
                 ticks="outside",
                 len=0.5,
+                x=0.9,
+                xref="paper", # Superpose the colobar with the map
             ),
             showscale=True,
             cmin=beta_ondergrsns - 1.5,
@@ -1068,6 +1094,8 @@ def add_colorscale_bar(
                 ticktext=["2", "3", str(round(beta_ondergrsns, 1)), "4", "5"],
                 ticks="outside",
                 len=0.5,
+                x=0.9,
+                xref="paper",  # Superpose the colobar with the map
             ),
             showscale=True,
             cmin=beta_ondergrsns - 1.5,