+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Added a button and name field to export the traject run json
 - Added a switch for the map of the program definition to display faalkans
+- Added a comparison map for the measures of 2 runs
 
 ### Changed
 - Map below reliability vs cost graph now displays the selected measure